script: Merge Window::load_url with ScriptThread::navigate
Both Window::load_url and ScriptThread::navigate implement parts of
the same navigate algorithm from the HTML spec
(https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigate).
ScriptThread::navigate had only one caller: Window::load_url.

This merges both methods into a single navigate function in
navigation.rs, which is the appropriate home for navigation logic.
All previous callers of Window::load_url now call navigation::navigate
directly.

Fixes: #43494
