+++ components/script/links.rs
@@ -21,6 +21,7 @@ use crate::dom::html::htmlformelement::HTMLFormElement;
 use crate::dom::html::htmllinkelement::HTMLLinkElement;
 use crate::dom::node::NodeTraits;
 use crate::dom::types::Element;
+use crate::navigation::navigate;
 use crate::script_runtime::CanGc;
 
 bitflags::bitflags! {
@@ -498,7 +499,7 @@ pub(crate) fn follow_hyperlink(
         let target = Trusted::new(target_window);
         let task = task!(navigate_follow_hyperlink: move |cx| {
             debug!("following hyperlink to {}", load_data.url);
-            target.root().load_url(history_handling, false, load_data, CanGc::from_cx(cx));
+            navigate(&target.root(), history_handling, false, load_data, CanGc::from_cx(cx));
         });
         target_document
             .owner_global()
