+++ components/script/navigation.rs
@@ -11,7 +11,7 @@ use std::cell::Cell;
 use content_security_policy::sandboxing_directive::SandboxingFlagSet;
 use crossbeam_channel::Sender;
 use embedder_traits::user_contents::UserContentManagerId;
-use embedder_traits::{Theme, ViewportDetails};
+use embedder_traits::{Theme, ViewportDetails, WebDriverLoadStatus};
 use http::header;
 use net_traits::policy_container::RequestPolicyContainer;
 use net_traits::request::{
@@ -26,11 +26,18 @@ use net_traits::{
 use script_traits::{DocumentActivity, NewPipelineInfo};
 use servo_base::cross_process_instant::CrossProcessInstant;
 use servo_base::id::{BrowsingContextId, PipelineId, WebViewId};
-use servo_constellation_traits::LoadData;
+use servo_constellation_traits::{
+    LoadData, LoadOrigin, NavigationHistoryBehavior, ScriptToConstellationMessage,
+};
 use servo_url::{ImmutableOrigin, MutableOrigin, ServoUrl};
+use url::Position;
 
+use crate::dom::bindings::refcounted::Trusted;
+use crate::dom::window::Window;
 use crate::fetch::FetchCanceller;
 use crate::messaging::MainThreadScriptMsg;
+use crate::script_runtime::CanGc;
+use crate::script_thread::ScriptThread;
 
 #[derive(Clone)]
 pub struct NavigationListener {
@@ -283,4 +290,133 @@ pub(crate) fn determine_the_origin(
 
     // Step 5. Return url's origin.
     MutableOrigin::new(url.origin())
+}
+
+/// <https://html.spec.whatwg.org/multipage/#navigate>
+pub(crate) fn navigate(
+    window: &Window,
+    history_handling: NavigationHistoryBehavior,
+    force_reload: bool,
+    load_data: LoadData,
+    can_gc: CanGc,
+) {
+    let doc = window.Document();
+
+    // Step 3. Let initiatorOriginSnapshot be sourceDocument's origin.
+    let initiator_origin_snapshot = &load_data.load_origin;
+
+    // TODO: Important re security. See https://github.com/servo/servo/issues/23373
+    // Step 5. check that the source browsing-context is "allowed to navigate" this window.
+
+    // Step 4 and 5
+    let pipeline_id = window.pipeline_id();
+    let window_proxy = window.window_proxy();
+    if let Some(active) = window_proxy.currently_active() {
+        if pipeline_id == active && doc.is_prompting_or_unloading() {
+            return;
+        }
+    }
+
+    // Step 23. Let unloadPromptCanceled be the result of checking if unloading
+    // is canceled for navigable's active document's inclusive descendant navigables.
+    if doc.check_if_unloading_is_cancelled(false, can_gc) {
+        // Step 12. If historyHandling is "auto", then:
+        let history_handling = if history_handling == NavigationHistoryBehavior::Auto {
+            // Step 12.1. If url equals navigable's active document's URL, and
+            // initiatorOriginSnapshot is same origin with targetNavigable's active document's
+            // origin, then set historyHandling to "replace".
+            //
+            // Note: `targetNavigable` is not actually defined in the spec, "active document" is
+            // assumed to be the correct reference based on WPT results
+            if let LoadOrigin::Script(initiator_origin) = initiator_origin_snapshot {
+                if load_data.url == doc.url() && initiator_origin.same_origin(&*doc.origin()) {
+                    NavigationHistoryBehavior::Replace
+                } else {
+                    // Step 12.2. Otherwise, set historyHandling to "push".
+                    NavigationHistoryBehavior::Push
+                }
+            } else {
+                // Step 12.2. Otherwise, set historyHandling to "push".
+                NavigationHistoryBehavior::Push
+            }
+        } else {
+            history_handling
+        };
+
+        // Step 13. If the navigation must be a replace given url and navigable's active
+        // document, then set historyHandling to "replace".
+        //
+        // Inlines implementation of https://html.spec.whatwg.org/multipage/#the-navigation-must-be-a-replace
+        let history_handling =
+            if load_data.url.scheme() == "javascript" || doc.is_initial_about_blank() {
+                NavigationHistoryBehavior::Replace
+            } else {
+                history_handling
+            };
+
+        // Step 14. If all of the following are true:
+        // > documentResource is null;
+        // > response is null;
+        if !force_reload
+            // > url equals navigable's active session history entry's URL with exclude fragments set to true; and
+            && load_data.url.as_url()[..Position::AfterQuery] ==
+                doc.url().as_url()[..Position::AfterQuery]
+            // > url's fragment is non-null,
+            && load_data.url.fragment().is_some()
+        {
+            // Step 14.1. Navigate to a fragment given navigable, url, historyHandling,
+            // userInvolvement, sourceElement, navigationAPIState, and navigationId.
+            let webdriver_sender = window.webdriver_load_status_sender();
+            if let Some(ref sender) = webdriver_sender {
+                let _ = sender.send(WebDriverLoadStatus::NavigationStart);
+            }
+            window.navigate_to_fragment(&load_data.url, history_handling);
+            // Step 14.2. Return.
+            if let Some(sender) = webdriver_sender {
+                let _ = sender.send(WebDriverLoadStatus::NavigationStop);
+            }
+            return;
+        }
+
+        // Step 15. If navigable's parent is non-null, then set navigable's is delaying load events to true.
+        let window_proxy = window.window_proxy();
+        if window_proxy.parent().is_some() {
+            window_proxy.start_delaying_load_events_mode();
+        }
+
+        if let Some(sender) = window.webdriver_load_status_sender() {
+            let _ = sender.send(WebDriverLoadStatus::NavigationStart);
+        }
+
+        // Step 13 of <https://html.spec.whatwg.org/multipage/#navigate>
+        let is_javascript = load_data.url.scheme() == "javascript";
+        if is_javascript {
+            let global = window.as_global_scope();
+            let trusted_global = Trusted::new(global);
+            let sender = global.script_to_constellation_chan().clone();
+            let mut load_data = load_data;
+            load_data.about_base_url = window.Document().about_base_url();
+            let task = task!(navigate_javascript: move |cx| {
+                // Important re security. See https://github.com/servo/servo/issues/23373
+                if trusted_global.root().is::<Window>() {
+                    let global = &trusted_global.root();
+                    if ScriptThread::navigate_to_javascript_url(cx, global, global, &mut load_data, None) {
+                        sender
+                            .send(ScriptToConstellationMessage::LoadUrl(load_data, history_handling))
+                            .unwrap();
+                    }
+                }
+            });
+            // Step 20 of <https://html.spec.whatwg.org/multipage/#navigate>
+            global
+                .task_manager()
+                .navigation_and_traversal_task_source()
+                .queue(task);
+        } else {
+            window.send_to_constellation(ScriptToConstellationMessage::LoadUrl(
+                load_data,
+                history_handling,
+            ));
+        }
+    };
 }
